Removed legacy Java 1.1 Code
Removed all references to legacy Java 1.1 code like `java.util.Vector' thus breaking source code compatibility in `jargs.gnu.CmdLineParser.getOptionValues'. Moreover replaced JUnit 3.x Testcases with JUnit 4.x annotated tests.

This effectivly makes JArgs a Java 1.5 compatible. No more unchecked warnings are emmited during compilation.
