+++ src/jargs/examples/gnu/AutoHelpParser.java
@@ -42,42 +42,42 @@ import java.util.List;
  * This example shows how to dynamically create basic output for a --help option. 
  */
 public class AutoHelpParser extends CmdLineParser {
-    List optionHelpStrings = new ArrayList();
 
-    public Option addHelp(Option option, String helpString) {
+    List<String> optionHelpStrings = new ArrayList<String>();
+
+    public <T> Option<T> addHelp(Option<T> option, String helpString) {
         optionHelpStrings.add(" -" + option.shortForm() + "/--" + option.longForm() + ": " + helpString);
         return option;
     }
     
     public void printUsage() {
         System.err.println("usage: prog [options]");
-        for (Iterator i = optionHelpStrings.iterator(); i.hasNext(); ) {
-            System.err.println(i.next());
+        for (String help : optionHelpStrings) {
+            System.err.println(help);
         }
     }
 
     public static void main( String[] args ) {
         AutoHelpParser parser = new AutoHelpParser();
-        CmdLineParser.Option verbose = parser.addHelp(
+        CmdLineParser.Option<Boolean> verbose = parser.addHelp(
                 parser.addBooleanOption('v', "verbose"),
                 "Print extra information");
-        CmdLineParser.Option size = parser.addHelp(
+        CmdLineParser.Option<Integer> size = parser.addHelp(
                 parser.addIntegerOption('s', "size"),
                 "The extent of the thing");
-        CmdLineParser.Option name = parser.addHelp(
+        CmdLineParser.Option<String> name = parser.addHelp(
                 parser.addStringOption('n', "name"),
                 "Name given to the widget");
-        CmdLineParser.Option fraction = parser.addHelp(
+        CmdLineParser.Option<Double> fraction = parser.addHelp(
                 parser.addDoubleOption('f', "fraction"),
                 "What percentage should be discarded");
-        CmdLineParser.Option help = parser.addHelp(
+        CmdLineParser.Option<Boolean> help = parser.addHelp(
                 parser.addBooleanOption('h', "help"),
                 "Show this help message");
 
         try {
             parser.parse(args);
-        }
-        catch ( CmdLineParser.OptionException e ) {
+        } catch ( CmdLineParser.OptionException e ) {
             System.err.println(e.getMessage());
             parser.printUsage();
             System.exit(2);
