+++ src/jargs/examples/gnu/CustomOptionTest.java
@@ -49,18 +49,20 @@ public class CustomOptionTest {
      * A custom type of command line option corresponding to a short
      * date value, e.g. .
      */
-    public static class ShortDateOption extends CmdLineParser.Option {
+    public static class ShortDateOption extends CmdLineParser.Option<Date> {
+
         public ShortDateOption( char shortForm, String longForm ) {
             super(shortForm, longForm, true);
         }
-        protected Object parseValue( String arg, Locale locale )
+
+        @Override
+        protected Date parseValue( String arg, Locale locale )
             throws CmdLineParser.IllegalOptionValueException {
             try {
                 DateFormat dateFormat =
                     DateFormat.getDateInstance(DateFormat.SHORT, locale);
                 return dateFormat.parse(arg);
-            }
-            catch (ParseException e) {
+            } catch (ParseException e) {
                 throw new CmdLineParser.IllegalOptionValueException(this, arg);
             }
         }
@@ -68,13 +70,12 @@ public class CustomOptionTest {
 
     public static void main( String[] args ) {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option date =
+        CmdLineParser.Option<Date> date =
             parser.addOption(new ShortDateOption('d', "date"));
 
         try {
             parser.parse(args);
-        }
-        catch ( CmdLineParser.OptionException e ) {
+        } catch ( CmdLineParser.OptionException e ) {
             System.err.println(e.getMessage());
             printUsage();
             System.exit(2);
@@ -83,7 +84,7 @@ public class CustomOptionTest {
         // Extract the values entered for the various options -- if the
         // options were not specified, the corresponding values will be
         // null.
-        Date dateValue = (Date)parser.getOptionValue(date);
+        Date dateValue = parser.getOptionValue(date);
 
         // For testing purposes, we just print out the option values
         System.out.println("date: " + dateValue);
