+++ src/jargs/test/gnu/CmdLineParserTestCase.java
@@ -33,28 +33,29 @@
 package jargs.test.gnu;
 
 import jargs.gnu.CmdLineParser;
+import jargs.gnu.CmdLineParser.Option;
 
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.Locale;
 
-import junit.framework.TestCase;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.fail;
+import org.junit.Test;
 
-public class CmdLineParserTestCase extends TestCase {
+public class CmdLineParserTestCase {
 
-    public CmdLineParserTestCase(String name) {
-        super(name);
-    }
-
+    @Test
     public void testStandardOptions() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
-        CmdLineParser.Option size = parser.addIntegerOption('s', "size");
-        CmdLineParser.Option name = parser.addStringOption('n', "name");
-        CmdLineParser.Option fraction = parser.addDoubleOption('f', "fraction");
-        CmdLineParser.Option missing = parser.addBooleanOption('m', "missing");
-        CmdLineParser.Option careful = parser.addBooleanOption("careful");
-        CmdLineParser.Option bignum = parser.addLongOption('b', "bignum");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
+        Option<Integer> size = parser.addIntegerOption('s', "size");
+        Option<String> name = parser.addStringOption('n', "name");
+        Option<Double> fraction = parser.addDoubleOption('f', "fraction");
+        Option<Boolean> missing = parser.addBooleanOption('m', "missing");
+        Option<Boolean> careful = parser.addBooleanOption("careful");
+        Option<Long> bignum = parser.addLongOption('b', "bignum");
         assertEquals(null, parser.getOptionValue(size));
         Long longValue = new Long(new Long(Integer.MAX_VALUE).longValue() + 1);
         parser.parse(new String[] { "-v", "--size=100", "-b",
@@ -62,32 +63,32 @@ public class CmdLineParserTestCase extends TestCase {
                 Locale.US);
         assertEquals(null, parser.getOptionValue(missing));
         assertEquals(Boolean.TRUE, parser.getOptionValue(verbose));
-        assertEquals(100, ((Integer) parser.getOptionValue(size)).intValue());
+        assertEquals(100, parser.getOptionValue(size).intValue());
         assertEquals("foo", parser.getOptionValue(name));
         assertEquals(longValue, parser.getOptionValue(bignum));
-        assertEquals(0.1, ((Double) parser.getOptionValue(fraction))
-                .doubleValue(), 0.1e-6);
+        assertEquals(0.1, parser.getOptionValue(fraction).doubleValue(), 0.1e-6);
         assertArrayEquals(new String[]{"rest"}, parser.getRemainingArgs());
     }
 
 
+    @Test
     public void testDefaults() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option boolean1 = parser.addBooleanOption("boolean1");
-        CmdLineParser.Option boolean2 = parser.addBooleanOption("boolean2");
-        CmdLineParser.Option boolean3 = parser.addBooleanOption("boolean3");
-        CmdLineParser.Option boolean4 = parser.addBooleanOption("boolean4");
-        CmdLineParser.Option boolean5 = parser.addBooleanOption("boolean5");
+        Option<Boolean> boolean1 = parser.addBooleanOption("boolean1");
+        Option<Boolean> boolean2 = parser.addBooleanOption("boolean2");
+        Option<Boolean> boolean3 = parser.addBooleanOption("boolean3");
+        Option<Boolean> boolean4 = parser.addBooleanOption("boolean4");
+        Option<Boolean> boolean5 = parser.addBooleanOption("boolean5");
 
-        CmdLineParser.Option int1 = parser.addIntegerOption("int1");
-        CmdLineParser.Option int2 = parser.addIntegerOption("int2");
-        CmdLineParser.Option int3 = parser.addIntegerOption("int3");
-        CmdLineParser.Option int4 = parser.addIntegerOption("int4");
+        Option<Integer> int1 = parser.addIntegerOption("int1");
+        Option<Integer> int2 = parser.addIntegerOption("int2");
+        Option<Integer> int3 = parser.addIntegerOption("int3");
+        Option<Integer> int4 = parser.addIntegerOption("int4");
 
-        CmdLineParser.Option string1 = parser.addStringOption("string1");
-        CmdLineParser.Option string2 = parser.addStringOption("string2");
-        CmdLineParser.Option string3 = parser.addStringOption("string3");
-        CmdLineParser.Option string4 = parser.addStringOption("string4");
+        Option<String> string1 = parser.addStringOption("string1");
+        Option<String> string2 = parser.addStringOption("string2");
+        Option<String> string3 = parser.addStringOption("string3");
+        Option<String> string4 = parser.addStringOption("string4");
 
         parser.parse(new String[] {
           "--boolean1", "--boolean2",
@@ -119,11 +120,12 @@ public class CmdLineParserTestCase extends TestCase {
     }
 
 
+    @Test
     public void testMultipleUses() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
-        CmdLineParser.Option foo = parser.addBooleanOption('f', "foo");
-        CmdLineParser.Option bar = parser.addBooleanOption('b', "bar");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
+        Option<Boolean> foo = parser.addBooleanOption('f', "foo");
+        Option<Boolean> bar = parser.addBooleanOption('b', "bar");
 
         parser.parse(new String[] {
           "--foo", "-v", "-v", "--verbose", "-v", "-b", "rest"
@@ -131,7 +133,7 @@ public class CmdLineParserTestCase extends TestCase {
 
         int verbosity = 0;
         while (true) {
-            Boolean b = (Boolean)parser.getOptionValue(verbose);
+            Boolean b = parser.getOptionValue(verbose);
 
             if (b == null) {
                 break;
@@ -150,11 +152,12 @@ public class CmdLineParserTestCase extends TestCase {
     }
 
 
+    @Test
     public void testCombinedFlags() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option alt = parser.addBooleanOption('a', "alt");
-        CmdLineParser.Option debug = parser.addBooleanOption('d', "debug");
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
+        Option<Boolean> alt = parser.addBooleanOption('a', "alt");
+        Option<Boolean> debug = parser.addBooleanOption('d', "debug");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
         parser.parse(new String[] {
           "-dv"
         });
@@ -165,12 +168,13 @@ public class CmdLineParserTestCase extends TestCase {
     }
 
 
+    @Test
     public void testExplictlyTerminatedOptions() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option alt = parser.addBooleanOption('a', "alt");
-        CmdLineParser.Option debug = parser.addBooleanOption('d', "debug");
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
-        CmdLineParser.Option fraction = parser.addDoubleOption('f', "fraction");
+        Option<Boolean> alt = parser.addBooleanOption('a', "alt");
+        Option<Boolean> debug = parser.addBooleanOption('d', "debug");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
+        Option<Double> fraction = parser.addDoubleOption('f', "fraction");
         parser.parse(new String[] {
           "-a", "hello", "-d", "-f", "10", "--", "goodbye", "-v", "welcome",
           "-f", "-10"
@@ -187,21 +191,22 @@ public class CmdLineParserTestCase extends TestCase {
     }
 
 
+    @Test
     public void testGetOptionValues() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
-        CmdLineParser.Option foo = parser.addBooleanOption('f', "foo");
-        CmdLineParser.Option bar = parser.addBooleanOption('b', "bar");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
+        Option<Boolean> foo = parser.addBooleanOption('f', "foo");
+        Option<Boolean> bar = parser.addBooleanOption('b', "bar");
 
         parser.parse(new String[] {
           "--foo", "-v", "-v", "--verbose", "-v", "-b", "rest"
         });
 
         int verbosity = 0;
-        Collection<Object> v = parser.getOptionValues(verbose);
-        Iterator<Object> it = v.iterator();
+        Collection<Boolean> v = parser.getOptionValues(verbose);
+        Iterator<Boolean> it = v.iterator();
         while (it.hasNext()) {
-            Boolean b = (Boolean)it.next();
+            Boolean b = it.next();
 
             if (b == Boolean.TRUE) {
                 verbosity++;
@@ -215,9 +220,10 @@ public class CmdLineParserTestCase extends TestCase {
     }
 
 
+    @Test
     public void testBadFormat() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option size = parser.addIntegerOption('s', "size");
+        Option<Integer> size = parser.addIntegerOption('s', "size");
         try {
             parser.parse(new String[] { "--size=blah" });
             fail("Expected IllegalOptionValueException");
@@ -226,29 +232,30 @@ public class CmdLineParserTestCase extends TestCase {
         }
     }
 
+    @Test
     public void testResetBetweenParse() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option verbose = parser.addBooleanOption('v', "verbose");
+        Option<Boolean> verbose = parser.addBooleanOption('v', "verbose");
         parser.parse(new String[] { "-v" });
         assertEquals(Boolean.TRUE, parser.getOptionValue(verbose));
         parser.parse(new String[] {});
         assertEquals(null, parser.getOptionValue(verbose));
     }
 
+    @Test
     public void testLocale() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option fraction = parser.addDoubleOption('f', "fraction");
+        Option<Double> fraction = parser.addDoubleOption('f', "fraction");
         parser.parse(new String[] { "--fraction=0.2" }, Locale.US);
-        assertEquals(0.2, ((Double) parser.getOptionValue(fraction))
-                .doubleValue(), 0.1e-6);
+        assertEquals(0.2, parser.getOptionValue(fraction).doubleValue(), 0.1e-6);
         parser.parse(new String[] { "--fraction=0,2" }, Locale.GERMANY);
-        assertEquals(0.2, ((Double) parser.getOptionValue(fraction))
-                .doubleValue(), 0.1e-6);
+        assertEquals(0.2, parser.getOptionValue(fraction).doubleValue(), 0.1e-6);
     }
 
+    @Test
     public void testDetachedOption() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option detached = new CmdLineParser.Option.BooleanOption(
+        Option<Boolean> detached = new CmdLineParser.Option.BooleanOption(
                 'v', "verbose");
         assertEquals(null, parser.getOptionValue(detached));
         try {
@@ -260,6 +267,7 @@ public class CmdLineParserTestCase extends TestCase {
         assertEquals(null, parser.getOptionValue(detached));
     }
 
+    @Test
     public void testMissingValueForStringOption() throws Exception {
         CmdLineParser parser = new CmdLineParser();
         parser.addBooleanOption('v', "verbose");
@@ -272,9 +280,10 @@ public class CmdLineParserTestCase extends TestCase {
         }
     }
 
+    @Test
     public void testWhitespaceValueForStringOption() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option opt = parser.addStringOption('o', "option");
+        Option<String> opt = parser.addStringOption('o', "option");
         parser.parse(new String[] {"-o", " "});
         assertEquals(" ", parser.getOptionValue(opt));
     }
