+++ src/jargs/test/gnu/CustomOptionTestCase.java
@@ -33,6 +33,7 @@
 package jargs.test.gnu;
 
 import jargs.gnu.CmdLineParser;
+import jargs.gnu.CmdLineParser.Option;
 
 import java.text.DateFormat;
 import java.text.ParseException;
@@ -40,38 +41,38 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
 
-import junit.framework.TestCase;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
+import org.junit.Test;
 
-public class CustomOptionTestCase extends TestCase {
+public class CustomOptionTestCase {
 
-    public CustomOptionTestCase(String name) {
-        super(name);
-    }
-
+    @Test
     public void testCustomOption() throws Exception {
         Calendar calendar = Calendar.getInstance();
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option date =
+        Option<Date> date =
             parser.addOption(new ShortDateOption('d', "date"));
 
         parser.parse(new String[]{"-d", "11/03/2003"}, Locale.UK);
-        Date d = (Date)parser.getOptionValue(date);
+        Date d = parser.getOptionValue(date);
         calendar.setTime(d);
         assertEquals(11,             calendar.get(Calendar.DAY_OF_MONTH));
         assertEquals(Calendar.MARCH, calendar.get(Calendar.MONTH));
         assertEquals(2003,           calendar.get(Calendar.YEAR));
 
         parser.parse(new String[]{"-d", "11/03/2003"}, Locale.US);
-        d = (Date)parser.getOptionValue(date);
+        d = parser.getOptionValue(date);
         calendar.setTime(d);
         assertEquals(3,                 calendar.get(Calendar.DAY_OF_MONTH));
         assertEquals(Calendar.NOVEMBER, calendar.get(Calendar.MONTH));
         assertEquals(2003,              calendar.get(Calendar.YEAR));
     }
 
+    @Test
     public void testIllegalCustomOption() throws Exception {
         CmdLineParser parser = new CmdLineParser();
-        CmdLineParser.Option date =
+        Option<Date> date =
             parser.addOption(new ShortDateOption('d', "date"));
         try {
             parser.parse(new String[]{"-d", "foobar"}, Locale.US);
@@ -82,11 +83,14 @@ public class CustomOptionTestCase extends TestCase {
         }
     }
 
-    public static class ShortDateOption extends CmdLineParser.Option {
+    public static class ShortDateOption extends CmdLineParser.Option<Date> {
+
         public ShortDateOption( char shortForm, String longForm ) {
             super(shortForm, longForm, true);
         }
-        protected Object parseValue( String arg, Locale locale )
+
+        @Override
+        protected Date parseValue( String arg, Locale locale )
             throws CmdLineParser.IllegalOptionValueException {
             try {
                 DateFormat dateFormat =
